Refactoring success screen
