+++ src/routes/types.routes.ts
@@ -18,7 +18,7 @@ export type RootStackParamList = {
 	WalletCreatedScreen: undefined;
 	WalletScreen: undefined;
 	WelcomeScreen: undefined; // undefined because you aren't passing any params to the home screen
-	DepositSuccessScreen: undefined;
+	DepositWithdrawalSuccessScreen: undefined;
 	BackupToICloudScreen: { missingPassword: boolean; walletId?: string | undefined; restoreBackups?: boolean };
 	OpenAave: undefined;
 	TopUpWaitScreen: undefined;
